requirements.txt
Flask==2.0.1
email-validator==1.1.3
Werkzeug==2.0.1
Jinja2==3.0.1
itsdangerous==2.0.1
MarkupSafe==2.0.1
click==8.0.1
python-dotenv==0.19.0
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.1
SQLAlchemy==1.4.23
Flask-Migrate==3.1.0
alembic==1.7.1
